Compra realizada con exito, falta conectarla con la API y enviarla

+++ database/migrations/2025_01_16_054249_create_orders_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
             $table->foreignId('bill_id');
             $table->foreignId('product_id');
             $table->integer('amount');
-            $table->integer('total_price');
+            $table->decimal('total_price');
 
             $table->foreign('bill_id')->references('id')->on('bills');
             $table->foreign('product_id')->references('id')->on('products');
